Create Module 1: Case Study - 1
Problem Statement:
Consider yourself to be Sam who is a data scientist. He has been invited as a
guest lecturer at a college to take an introductory session on Python.
Tasks To Be Performed:
1. Create a list containing squares of numbers from 1 to 10 (HINT: use List
Comprehension).
2. Write a function to check if the year number is a leap year.
3. Write a function to take an array and return another array that contains the
members of the first array that are even.
4. Write a function that takes 2 arrays and prints the members of the first
array that are present in the second array. (HINT: use Membership
Comprehension)